Extract useWebviewReady hook (oh-tab-ihgw)
